+++ pkg/service/admin/filesystem/filesystem-service.go
@@ -6,6 +6,7 @@ import (
 
 	adminv2 "github.com/metal-stack/api/go/metalstack/admin/v2"
 	"github.com/metal-stack/api/go/metalstack/admin/v2/adminv2connect"
+	"github.com/metal-stack/metal-apiserver/pkg/errorutil"
 	"github.com/metal-stack/metal-apiserver/pkg/repository"
 )
 
@@ -54,4 +55,22 @@ func (f *filesystemServiceServer) Update(ctx context.Context, rq *adminv2.Filesy
 	}
 
 	return &adminv2.FilesystemServiceUpdateResponse{FilesystemLayout: fsl}, nil
+}
+func (f *filesystemServiceServer) Match(ctx context.Context, req *adminv2.FilesystemServiceMatchRequest) (*adminv2.FilesystemServiceMatchResponse, error) {
+	switch match := req.Match.(type) {
+	case *adminv2.FilesystemServiceMatchRequest_SizeAndImage:
+		fsl, err := f.repo.FilesystemLayout().AdditionalMethods().Try(ctx, match.SizeAndImage)
+		if err != nil {
+			return nil, err
+		}
+		return &adminv2.FilesystemServiceMatchResponse{FilesystemLayout: fsl}, nil
+	case *adminv2.FilesystemServiceMatchRequest_MachineAndFilesystemlayout:
+		fsl, err := f.repo.FilesystemLayout().AdditionalMethods().Match(ctx, match.MachineAndFilesystemlayout)
+		if err != nil {
+			return nil, err
+		}
+		return &adminv2.FilesystemServiceMatchResponse{FilesystemLayout: fsl}, nil
+	default:
+		return nil, errorutil.InvalidArgument("given matchtype %T is unsupported", match)
+	}
 }
